fix(predictor/real-bid): 투찰가 정수 정합 — 최종 예측가 원 단위 보장 + 등록 bid_amount 정수 강제 (#203)
* fix(predictor): 예측 최종 투찰가 원 단위 정수 양자화 보장

시스템이 내놓는 투찰가 = KONEPS에 그대로 제출 가능한 정수(원) 계약을 강제한다.
기존 apply_bid_price_granularity는 granularity≤1 또는 예산<min_budget이면 전부
스킵되어 소수 투찰가(예: 785.2원)가 최종 출력으로 노출될 수 있었다.

- budget<=0만 조기 반환 유지(predicted_bid_rate = price/budget 나눗셈 분모).
- 그 외에는 effective unit = granularity(적용 조건: granularity>1 AND 예산≥min_budget)
  else 1(원 단위)로 양자화. 기존 round_price_to_granularity 기계 재사용(floor/ceiling
  클램프·rounding_mode 동일).
- 메타 bid_price_granularity에 실제 적용 unit(1 또는 granularity) 정직 기록.
- 기존 granularity 적용 케이스(예산≥min_budget)의 출력은 불변 — 동작 변경은 오직
  "기존에 스킵되던 케이스가 원 단위 정수가 되는 것"뿐(guardrail 골든 스위트 무변).
- 테스트: 예산<min_budget→785.2→785, granularity=0→원 단위, budget<=0 무변경,
  floor 보호 unit=1 유지, candidate 경로 포함 값테이블 추가.


* fix(real-bid): 실투찰 등록 bid_amount 정수 강제 + 타입 동기화

실투찰 사후분석에서 소수 투찰가가 등록되어 노출값과 불일치가 발생했다. 등록
경계에서 정수(원)만 허용해 KONEPS 제출값과의 정합을 코드로 강제한다.

- RealBidRecordRequest.bid_amount를 int(gt=0)로 변경("실제 제출한 투찰가(원, 정수)").
  pydantic v2 lax: 785.2 같은 소수는 422, 77529790.0 같은 정수값 float은 통과.
- RealBidTrackService.record_real_bid의 bid_amount 힌트 int 정합화(저장은 기존 Float
  컬럼 그대로 float(bid_amount) — 마이그레이션 없음). floor_rate는 비율이라 float 유지.
- 테스트: 소수 422, 정수값 float 통과·저장값 정수 추가.
- OpenAPI: bid_amount number→integer. sync-types로 openapi.d.ts 재생성(TS는 integer도
  number 매핑이라 타입 불변, @description만 갱신 — 프론트 호환 깨짐 없음).


---------

+++ app/services/real_bid_track.py
@@ -64,7 +64,7 @@ class RealBidTrackService:
         *,
         operator: User,
         project: Project,
-        bid_amount: float,
+        bid_amount: int,
         floor_rate: float | None = None,
         submitted_at: datetime | None = None,
         note: str | None = None,
